Bugs/125/requirements.txt
numpy==2.1.0
Pillow==10.4.0
tensorflow==2.15.1
tqdm==4.66.4
